feature: use langchain's react agent and new architecture
