+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "matplotlib>=3.10.7",
     "numpy>=2.3.4",
     "plotly>=6.5.0",
+    "markdown>=3.10",
 ]
 
 [project.optional-dependencies]
