+++ src/vitess_ai/server/streaming/__init__.py
@@ -1,0 +1,25 @@
+"""
+Stream handlers for processing LangGraph streaming events.
+
+This module provides handlers for different stream modes (updates, messages, custom)
+to process and convert LangGraph streaming events into chat messages.
+"""
+
+from vitess_ai.server.streaming.processor import StreamEventProcessor
+from vitess_ai.server.streaming.handlers import (
+    UpdatesStreamHandler,
+    MessagesStreamHandler,
+    CustomStreamHandler,
+)
+from vitess_ai.server.streaming.message_processor import MessageProcessor
+from vitess_ai.server.streaming.deduplication import get_message_identifier
+
+__all__ = [
+    "StreamEventProcessor",
+    "UpdatesStreamHandler",
+    "MessagesStreamHandler",
+    "CustomStreamHandler",
+    "MessageProcessor",
+    "get_message_identifier",
+]
+
