+++ src/vitess_ai/server_agents/writeout_module_agent.py
@@ -6,13 +6,11 @@ architecture, enabling unified state management and
 centralized interrupt handling.
 """
 
-from typing import Type
 from vitess_ai.server_agents.base_module_agent import BaseModuleAgent
 from vitess_ai.prompts.writeout_module import WRITEOUT_AGENT_WELCOME, WRITEOUT_AGENT_DEFAULT_PROMPT, WRITEOUT_AGENT_CUSTOM_PROMPT
-from vitess_ai.schema.writeout_module import InitialResponseWriteout
 
 
-class WriteoutModuleAgent(BaseModuleAgent[InitialResponseWriteout]):
+class WriteoutModuleAgent(BaseModuleAgent):
     """
     Writeout module agent.
     
@@ -45,47 +43,9 @@ class WriteoutModuleAgent(BaseModuleAgent[InitialResponseWriteout]):
         """Custom prompt for the writeout module"""
         return WRITEOUT_AGENT_CUSTOM_PROMPT
     
-    def get_initial_response_schema(self) -> Type[InitialResponseWriteout]:
-        """Return the schema for initial response parsing"""
-        return InitialResponseWriteout
-    
     def get_result_key(self) -> str:
         """Return the key name for storing results"""
         return "writeout_params"
-    
-    def get_default_setup_message(self) -> str:
-        """Message shown when user chooses default setup"""
-        return """
-        You have chosen the default setup configuration for writeout parameters. 
-        We will use optimal default values for most output settings. 
-        You'll only need to specify essential parameters that require manual input.
-        
-        Default writeout parameters include:
-        - Standard data output formats (Nexus, ASCII)
-        - Typical file naming conventions
-        - Common data compression settings
-        - Standard metadata inclusion options
-        
-        The system will guide you through any parameters that need customization.
-        """
-    
-    def get_customize_setup_message(self) -> str:
-        """Message shown when user chooses customization"""
-        return """
-        You have chosen the customize configuration for writeout parameters. 
-        I'll help you configure the output settings step by step. 
-        We'll go through each parameter category and you can choose which ones 
-        to modify from the defaults.
-        
-        Configuration categories include:
-        - Data output formats (Nexus, ASCII, binary, etc.)
-        - File naming and organization
-        - Data compression and storage options
-        - Metadata and annotation settings
-        - Performance and optimization parameters
-        
-        Let's start with the data output format configuration.
-        """
     
     def get_completion_message(self, state: dict = None) -> str:
         """Message shown on successful completion"""
